Corrigindo bug

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -15,7 +15,7 @@ public class Calculadora {
 
         int soma = soma(a, b);
         int subtracao = subtracao(a, b);
-        int divisao = divisao(a, b);
+        double divisao = divisao(a, b);
         int multiplicacao = multiplicacao(a, b);
 
         System.out.println("Soma: " + soma);
@@ -34,7 +34,7 @@ public class Calculadora {
         return a - b;
     }
 
-    public static int divisao (int a, int b){
+    public static double divisao (double a, double b){
         return a / b;
     }
